requirements.txt
Flask==2.0.3
paho-mqtt==1.6.1